Refatoração de model

+++ src/main/java/com/leonardo/mangareader/repositories/ChapterRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.leonardo.mangareader.models.Chapter;
 
 @Repository
-public interface ChapterRepository extends JpaRepository<Chapter, Integer>{
+public interface ChapterRepository extends JpaRepository<Chapter, Long>{
 
     @Query("SELECT chapter FROM Chapter chapter WHERE chapter.url = ?1")
     public Optional<Chapter> findByUrl(String url);
