+++ src/main/java/com/leonardo/mangareader/repositories/MangaRepository.java
@@ -7,7 +7,7 @@ import com.leonardo.mangareader.models.Manga;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface MangaRepository extends JpaRepository<Manga, Integer>{
+public interface MangaRepository extends JpaRepository<Manga, Long>{
     
     @Query("SELECT manga FROM Manga manga WHERE manga.url = ?1")
     public Optional<Manga> findByUrl(String url);
